Add notificiations and better tab management
The pasted plugin will now notify the user that content has been added
to the paste buffer and open a tab at index 0 that is NOT active. This
ensures the users is not navigated away from the content they're working
on while also making it possible to rapidly share links to raw content
with others.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>